fix(iOS, FullWindowOverlay): Move show logic to didMoveToWindow (#3379)
## Description

In a multi-window/multi-scene app setup, the current logic for
determining the window uses the `RCTKeyWindow`:
https://github.com/facebook/react-native/blob/main/packages/react-native/React/Base/RCTUtils.mm#L615-L618

The logic selects the first scene that is in the foreground and returns
its key window. However, in our case, all scenes are marked as
foreground, so the utility may return a window belonging to a different
scene than the one associated with the component.

Additionally, due to the current implementation, the `show` method might
be called before the view is attached to any window. As a result,
`self.window` is `nil`, which prevents `FullWindowOverlay` from
initializing correctly.

This also caused a separate issue when the scene is running in the
background, where the component fails to mount due to the inability to
resolve the correct window context.

To ensure the correct window is always used, we now move the logic for
showing the overlay into `didMoveToWindow`. At this point, `self.window`
is guaranteed to be valid.

Fixes
https://github.com/software-mansion/react-native-screens/issues/3372

## Changes

- Added `didMoveToWindow` for reading the proper window connected with
the scene/application.

## Screenshots / GIFs

Here you can add screenshots / GIFs documenting your change.

You can add before / after section if you're changing some behavior.

### Before


https://github.com/user-attachments/assets/8ee8c05c-f906-4e89-a322-d21c78f63118

### After

|  | Single Window | Multiple Windows |
|-|----------------|-------------------|
| Paper |
https://github.com/user-attachments/assets/6d5e2cf6-5c79-4fa8-b6bf-441a15f6c74f
|
https://github.com/user-attachments/assets/f27080a8-e902-4d01-8eb5-5c9ba2e71c33
|
| Fabric |
https://github.com/user-attachments/assets/0ef41ea9-5e8a-487f-824d-68a52027e09e
|
https://github.com/user-attachments/assets/0f850118-4366-4f00-aa25-a9a9242ffed8
|

## Test code and steps to reproduce

Test3379 in the following configurations:
- [x] FabricExample - ApplicationDelegate
- [x] FabricExample - SceneDelegate
- [x] Example - ApplicationDelegate
- [x] Example - SceneDelegate

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes