chore: remove not necessary workaround for ScrollView from RN 76 (#2350)
## Description

<!--
Description and motivation for this PR.

Include Fixes #<number> if this is fixing some issue.

Fixes # .
-->

Due to fixes in RN 0.76
(https://github.com/facebook/react-native/commit/b98b9f1fa7717283f368eb182a51d971b8776c80
and
https://github.com/facebook/react-native/commit/c6f32828b9487381dab27f645aedcdbae9dcbc7e)
our workaround seems to not be necessary anymore. Would be nice to check
if it works in RN 0.76 on both archs before merging it 😄

---------

Co-authored-by: Kacper Kafara <kacperkafara@gmail.com>