fix(Android): layout shift on orientation change (#3295)
Fixes #2933

# Description

This PR fixes the bug on android that results in Screen orientation
being calculated incorrectly for some time after the physical device
orientation has changed. This bug seems to happen due to asynchronous
nature of Screen state update emitted to react from native layer. The
state contains a size that is set on the Screen frame.

The bug is more visible on slower devices, because with longer
processing of consecutive commits, the delay is longer.

It can be noticed on a trivial application (some text + different
background color for screen and content view) with some heavy
computations artificially added to JS.

### More detailed description:
In `RNSScreenComponentDescriptor`, on Android specifically, we call
`setSize()` on the Screen view to hardcode its dimensions. This is
needed to account for screen header size, of which React has no idea. It
is however the core reason for the bug. The size which we set is
outdated. Only after the asynchronous state update from native side is
processed, it is corrected. This can take some time. If we comment the
code for handling the header size, this problem disappears. But we
cannot.

Since the code is staying here, we need to work with it. We agreed on
the special value of frameSize = { 0, 0 }, for which we set YogaNode
size to Undefined. This forces Yoga to recalculate the layout. It can do
so because the parent view is correctly sized at this point (no calls to
`setSize()` with outdated values). We cannot do this manually, because
the descriptor doesn't have access to parent nodes. Using the commit
hook, we detect the orientation change (it can compare old and new
RootShadowNode, which have the correct dimensions) and set the frameSize
for the ShadowNode which is passed to the descriptor afterwards.

Another problem is that the hook detects the change only once, for one
instance of the shadow node. There could be more instances processed
before the state is corrected. Luckily, React uses the same state
instance for each shadow node instance between the updates. We take
advantage of that, once reset to 0, the frameSize stays that way until
some other updates changes the instance. Since we only store the screen
offset and frameSize in the state, this works good enough. **But it's
still pretty hacky**.

# Changes

Description above. A new feature flag is added. When set, we register a
new commit hook that checks if screen orientation changed & sets
frameSize to 0. It is detected in Screen descriptor and Yoga is forced
to recalculate the layout.

# Before / After

| before | after |
| --- | --- |
| <video
src="https://github.com/user-attachments/assets/8ff1187c-1ac3-4a15-8ccf-f751c692dc44"
/> | <video
src="https://github.com/user-attachments/assets/d8ebdbb4-4500-4968-98ea-75b475041b76"
/> |

# Testing
Use Test2933. You can take a look at the "checkerboard" test from
react-native-reanimated.

---------

Co-authored-by: Andrei Calazans <andreixoc@hotmail.com>