feat: Remove native-stack v5 from screens repo (#3433)
## Description

This PR removes native-stack directory from screens repo. This dir held
the code for native-stack v5, which has been moved to react-navigation
and is v7, v8 now. The rest of the deprecated imports are removed and
the unused code is deleted.

Closes
https://github.com/software-mansion/react-native-screens-labs/issues/589

## Changes

Removed native-stack/*, fixed imports.


## Test code and steps to reproduce

Check if the library still builds successfully.