refactor: Remove codegen-related deep imports (#3458)
## Description

This PR initiates the migration towards the stable JS APIs introduced in
React Native 0.80+, as described in the official post
https://reactnative.dev/blog/2025/06/12/moving-towards-a-stable-javascript-api.

> [!NOTE]  
> As RN 0.79 does not export the required codegen bindings, this PR
should be merged **only after**
https://github.com/software-mansion/react-native-screens/pull/3457 is
merged.

> [!NOTE]
> There is still one remaining deep import
(`react-native/Libraries/ReactNative/AppContainer`) used in
`DebugContainer`. This relies on private APIs with a complex structure
(e.g., Inspector) that do not have a feasible workaround at this point.
The further investigation and potential resolution should be handled in
a follow-up PR to avoid mixing it with codegen-related changes made in
this one. -
https://github.com/software-mansion/react-native-screens-labs/issues/703
>
> EDIT: After checking, we're following the same approach as Modal
implementation in Core, therefore I asked Meta team whether it'd be
possible to expose `AppContainer` publicly:
https://github.com/react-native-community/discussions-and-proposals/discussions/893#discussioncomment-15195267

```js
 WARN  Deep imports from the 'react-native' package are deprecated ('react-native/Libraries/ReactNative/AppContainer'). Source: /Users/tomaszboron/react-native-screens/src/components/DebugContainer.tsx 5:0
```

Fixes
https://github.com/software-mansion/react-native-screens-labs/issues/378

## Changes

- Replaced deep imports for codegen with public RN APIs

## Test code and steps to reproduce

Launched FabricExample on both android and iOS. Tested backward
compatibility with RN 0.81 test app.

Listed all occurencies of `react-native/`

```bash
➜  pwd
/Users/tomaszboron/react-native-screens/src
➜  grep -r "react-native/" .
./components/DebugContainer.tsx:import AppContainer from 'react-native/Libraries/ReactNative/AppContainer';
./components/Screen.tsx:// react-native/Libraries/Components/View/ReactNativeViewViewConfig.js
```

Both are okay:
1. AppContainer, left because lack of the public API
2. Just a comment inside the component, not an import

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes