chore: Sync react-navigation with 7.x branch (#3468)
## Description

Including a fix for the TVOS focus issue on iOS.

Reversing the order in the view hierarchy introduced in this commit
https://github.com/react-navigation/react-navigation/commit/b068c1a1c075f85d54ff4e297ae1df312d3fc0f9
, resulted in the regression, because TV focus was operating in the
scope of this `View` with `absoluteFill` style, preventing all
interactive components inside children from catching focus. Then, the
hierarchy has changed and content is on the top layer for which focus
search operates.

Fixes
https://github.com/software-mansion/react-native-screens-labs/issues/708