Make retrieval of `UIManager` from `ContextContainer` a *little* bit more safe
Wanted to avoid more places where we can crash the app.
Unfortunately I can not make it fully safe since calling
`FabricUIManger->getBinding()` will still crash if the `mBinding`
stored on Java side is null.

In the future we could consider either sending a patch to core to
prevent the crash or research whether it is possible to access the field
manually.