Reimplement StackContainer
The original implementation is hard for me to follow. Moreover it has
been based upon old operation model based on `maxLifecycleState`.

New operation model introduced in RFC-0591 requires changes in
implementation of stack operations, therefore I took this opportunity
and reimplemented all of it.

The implementation is not completed yet (nor was the previous one).
It introduces single `StackContainer` component that should be used
to render the new stack.

While implementing I noticed a couple of problems introduced by
me in RFC-0591. First - there is no easy way to get the navigation state
in the JS layer. This is especially troublesome when implementing `pop`
operation, as it is not clear what screen to dismiss.

The exact problem lies in fact that in current model we allow for
rendering array of `Stack.Screen`s with some in `detached` mode.
When already rendered `Screen` changes its mode to `attached`,
it becomes a top screen, but it's index in the state array does not
change. Therefore, we can not really tell which screen is on top.

`Pop` operation is one problem, but another is simply dumping current
navigation state e.g. for telemetry purposes / saving app state
before restart or something. I'm sure there are many use cases for
reading current navigation state.