Fix typing for SearchBar ref
It seems that in React 19.0 if you pass `null` as the initArg
to the `React.useRef` then nullable type will be inferred anyway.
So the `searchBarRef` is of type `React.RefObject<SearchBarCommands |
null>` and we pass it to ref whose type is determined by
`NativeSearchBar` type definition, which assumes
`React.RefObject<SearchBarCommands>`, hence the error.

I believe applying this patch shouldn't be treated as breaking change,
but rather as a type fix. For anyone typechecking / using this ref
it must have been annoying. The error would have to be either
suppressed or ignored.