Pass info on `activityMode` change from StackScreen to StackHost
Before this change, the information has been passed from
`RNSStackScreenComponentView` to `StackController` via
`StackScreenController`. This is not compliant with RFC-0753.

The controllers should not be aware of activity mode changes. The
information should be passed directly to `StackHost` and there request
to container should be dispatched. This is exactly what this commit
does.